# docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"SettingResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"value": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"success",
					"value"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SettingActionResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SettingValueDto": {
				"properties": {
					"value": {
						"type": "string"
					}
				},
				"required": [
					"value"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"PaymentRecordDto": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"sessionId": {
						"type": "number",
						"format": "double"
					},
					"consoleId": {
						"type": "number",
						"format": "double"
					},
					"orderId": {
						"type": "string"
					},
					"paymentMethodId": {
						"type": "string"
					},
					"paymentMethodName": {
						"type": "string"
					},
					"amount": {
						"type": "number",
						"format": "double"
					},
					"isCash": {
						"type": "boolean"
					},
					"staff": {
						"type": "string"
					},
					"notes": {
						"type": "string"
					},
					"createdAt": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					}
				},
				"required": [
					"paymentMethodId",
					"paymentMethodName",
					"amount",
					"isCash",
					"staff"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PaymentListResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/PaymentRecordDto"
						},
						"type": "array"
					}
				},
				"required": [
					"success",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PaymentSplitDto": {
				"properties": {
					"paymentMethodId": {
						"type": "string"
					},
					"amount": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"paymentMethodId",
					"amount"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ProcessPaymentDto": {
				"properties": {
					"sessionId": {
						"type": "number",
						"format": "double"
					},
					"consoleId": {
						"type": "number",
						"format": "double"
					},
					"orderId": {
						"type": "string"
					},
					"payments": {
						"items": {
							"$ref": "#/components/schemas/PaymentSplitDto"
						},
						"type": "array"
					},
					"staff": {
						"type": "string"
					},
					"notes": {
						"type": "string"
					}
				},
				"required": [
					"payments"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"PaymentMethodDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"nameAr": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"isCash": {
						"type": "boolean"
					},
					"isProtected": {
						"type": "boolean"
					},
					"isActive": {
						"type": "boolean"
					},
					"displayOrder": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"name",
					"nameAr",
					"type",
					"isCash",
					"isProtected",
					"isActive",
					"displayOrder"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PaymentMethodListResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/PaymentMethodDto"
						},
						"type": "array"
					}
				},
				"required": [
					"success",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PaymentMethodResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"$ref": "#/components/schemas/PaymentMethodDto"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreatePaymentMethodDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"nameAr": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"isCash": {
						"type": "boolean"
					},
					"displayOrder": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"nameAr"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdatePaymentMethodDto": {
				"properties": {
					"name": {
						"type": "string"
					},
					"nameAr": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"isCash": {
						"type": "boolean"
					},
					"isActive": {
						"type": "boolean"
					},
					"displayOrder": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "playstation-system-backend",
		"version": "1.0.0",
		"contact": {}
	},
	"paths": {
...
		"/api/v1/settings/{key}": {
			"get": {
				"operationId": "GetSetting",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SettingResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Settings"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "key",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "SetSetting",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SettingActionResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Settings"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "key",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SettingValueDto"
							}
						}
					}
				}
			},
			"put": {
				"operationId": "UpdateSetting",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SettingActionResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Settings"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "key",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SettingValueDto"
							}
						}
					}
				}
			}
		},
...
		"/api/v1/payments": {
			"get": {
				"operationId": "GetPayments",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaymentListResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Payments"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "sessionId",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "consoleId",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "orderId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "isCash",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					}
				]
			},
			"post": {
				"operationId": "ProcessPayments",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaymentListResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Payments"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ProcessPaymentDto"
							}
						}
					}
				}
			}
		},
...
		"/api/v1/payment-methods": {
			"get": {
				"operationId": "GetAllPaymentMethods",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaymentMethodListResponse"
								}
							}
						}
					}
				},
				"tags": [
					"PaymentMethods"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreatePaymentMethod",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaymentMethodResponse"
								}
							}
						}
					}
				},
				"tags": [
					"PaymentMethods"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreatePaymentMethodDto"
							}
						}
					}
				}
			}
		},
		"/api/v1/payment-methods/{id}": {
			"get": {
				"operationId": "GetPaymentMethodById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaymentMethodResponse"
								}
							}
						}
					}
				},
				"tags": [
					"PaymentMethods"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdatePaymentMethod",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaymentMethodResponse"
								}
							}
						}
					}
				},
				"tags": [
					"PaymentMethods"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdatePaymentMethodDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeletePaymentMethod",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"message",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"PaymentMethods"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}